Add basic input validation and accessibility
User has to put something in the name fields, the phone number has to be
format xxx-xxx-xxxx, and the password has to be at least five characters
long.

Will do a better accessibility check later, but images have alt text and
fields are marked "required" + have straightforward labels and
placeholders.