Perfect form input fields
Inputs all have a 2px outline when not static, with appropriate gold,
eggplant, and bright red outline colors for focused, valid, and invalid
entries respectively. Red outline comes with a pink background, which
may be a bit much, but when you refocus on it it does revert to the
focus gold outline + white background + gentle shadow.

Images are now in an images folder for the sake of general tidiness.